Manage arm creation via TaskArm object
Previously arm-related aspects were handled by the Slurm object. This is
a bit illogical especially if we want the needed status to be per-arm
rather than per-task.

This commit introduces a new object called TaskArm
that is like a Task object, except that it is for a specific arm.
Generation of the arms is now handled by the Task object.