Neater, more robust cat_df
This version pads out short cells so that all the columns will always
line up. It does this while retaining ansi colour codes, but doesn't
count them in the number of characters that are in each string.

It also prevents overrunning of the table if the table is too wide for
the available terminal width, and inserts an elipsis to represent that
the output has been prematurely curtailed.

Also adds the get_terminal_width() wrapper function