Split prereq_ids into two function to account for dummies better
Dummy tasks introduce recursion into identifying prerequisites, and
the transposition then caused problems. To solve this, separate the
recursive and post-processing parts of the task up, so that the
tranposition only occurs once.

To simplify things, I now special-case the single dummy prereq, and
warn that more complex cases are not accounted for at the moment